feat(services/gcs): support sink (#2576)

+++ core/src/services/gcs/core.rs
@@ -207,7 +207,7 @@ impl GcsCore {
     pub fn gcs_insert_object_request(
         &self,
         path: &str,
-        size: Option<usize>,
+        size: Option<u64>,
         content_type: Option<&str>,
         body: AsyncBody,
     ) -> Result<Request<AsyncBody>> {
